# api/swagger/swagger.yaml
definitions:
  api.AddSongRequest:
    properties:
      group:
        example: Muse
        type: string
      song:
        example: Supermassive Black Hole
        type: string
    type: object
  api.PatchRequest:
    properties:
      group:
        example: Rick Astley
        type: string
      link:
        example: https://www.youtube.com/watch?v=dQw4w9WgXcQ
        type: string
      releaseDate:
        example: 24.07.1987
        type: string
      song:
        example: Never Gonna Give You Up
        type: string
      text:
        example: |-
          We’re no strangers to love
          You know the rules and so do I
          A full commitment’s what I’m thinking of
          You wouldn’t get this from any other guy

          I just wanna tell you how I’m feeling
          Gotta make you understand

          Never gonna give you up
          Never gonna let you down
          Never gonna run around and desert you
          Never gonna make you cry
          Never gonna say goodbye
          Never gonna tell a lie and hurt you

          We’ve known each other for so long
          Your heart’s been aching but you’re too shy to say it
          Inside we both know what’s been going on
          We know the game and we’re gonna play it

          And if you ask me how I’m feeling
          Don’t tell me you’re too blind to see

          Never gonna give you up
          Never gonna let you down
          Never gonna run around and desert you
          Never gonna make you cry
          Never gonna say goodbye
          Never gonna tell a lie and hurt you
        type: string
    type: object
  api.Song:
    properties:
      group:
        example: Muse
        type: string
      id:
        example: uuid
        type: string
      link:
        example: https://www.youtube.com/watch?v=Xsp3_a-PMTw
        type: string
      releaseDate:
        example: 16.07.2006
        type: string
      song:
        example: Supermassive Black Hole
        type: string
      text:
        example: |-
          oh baby, don't you know I suffer?
          Ooh baby, can you hear me moan?
          You caught me under false pretenses
          How long before you let me go?

          Ooh
          You set my soul alight
          Ooh
          You set my soul alight
        type: string
    type: object
  api.Text:
    properties:
      verse:
        example: |
          Never gonna give you up
        type: string
      verse_number:
        example: 1
        type: integer
    type: object
  rest.APIError:
    properties:
      msg:
        example: invalid ID
        type: string
      status_code:
        example: 400
        type: integer
    type: object
info:
  contact: {}
  title: Music Library API
  version: "1.0"
paths:
  /api/songs:
    get:
      description: Prints List of songs from library.
      parameters:
      - description: Page number. Default 1
        in: query
        name: page
        type: integer
      - description: Filter By {}. Default by 'creation time'
        in: query
        name: filter
        type: string
      - description: Order By 'asc' or 'desc'. Default 'asc'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Song'
            type: array
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      summary: Get Songs
      tags:
      - Songs
    post:
      consumes:
      - application/json
      description: Enriches song with additional inforamtion, then adds song to Library.
        If song inforamtion can't be enriched, error is shown
      parameters:
      - description: Band and Song names
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/api.AddSongRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Song'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.APIError'
        "404":
          description: Not Found
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      summary: Upload Song To Library
      tags:
      - Songs
  /api/songs/{id}:
    delete:
      description: Deletes song from Library
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      summary: Delete Song
      tags:
      - Songs
    patch:
      consumes:
      - application/json
      description: Updates song information in Library. If without body 204 code returned
        with no changes done
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: string
      - description: Change song info
        in: body
        name: group
        schema:
          $ref: '#/definitions/api.PatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Song'
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.APIError'
        "404":
          description: Not Found
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      summary: Update Song Info
      tags:
      - Songs
  /api/songs/{id}/text:
    get:
      description: Prints text with verse number
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: string
      - description: Verse number. If not provided prints full lirics
        in: query
        name: verse
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Text'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.APIError'
        "404":
          description: Not Found
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.APIError'
      summary: Get Song Text
      tags:
      - Songs
  /healthz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: object
      summary: Check API Availability
      tags:
      - Status
swagger: "2.0"